basic layout follows navigation plan
