+++ app/src/main/java/nyc/c4q/helenchan/makinghistory/MainActivity.java
@@ -1,98 +1,38 @@
 package nyc.c4q.helenchan.makinghistory;
 
+import android.content.Intent;
 import android.os.Bundle;
-import android.support.annotation.NonNull;
 import android.support.annotation.Nullable;
 import android.support.v7.app.AppCompatActivity;
-import android.widget.Toast;
+import android.view.View;
+import android.widget.Button;
+import android.widget.TextView;
 
-import com.google.android.gms.common.ConnectionResult;
-import com.google.android.gms.common.GoogleApiAvailability;
-import com.google.android.gms.common.api.GoogleApiClient;
-import com.google.android.gms.location.LocationServices;
-import com.google.android.gms.maps.CameraUpdateFactory;
-import com.google.android.gms.maps.GoogleMap;
-import com.google.android.gms.maps.OnMapReadyCallback;
-import com.google.android.gms.maps.SupportMapFragment;
-import com.google.android.gms.maps.model.LatLng;
-import com.google.android.gms.maps.model.MarkerOptions;
-import com.google.firebase.database.DatabaseReference;
-import com.google.firebase.database.FirebaseDatabase;
+public class MainActivity extends AppCompatActivity implements View.OnClickListener {
 
-public class MainActivity extends AppCompatActivity implements OnMapReadyCallback, GoogleApiClient.ConnectionCallbacks, GoogleApiClient.OnConnectionFailedListener {
-
-    private static final String TAG = "Main Activity";
-    GoogleMap mMap;
-    private final static int PLAY_SERVICES_RESOLUTION_REQUEST = 9000;
-
-    private GoogleApiClient mLocationClient;
-
-
-    private FirebaseDatabase mFirebaseDatabase;
-    private DatabaseReference mFirebasseRef;
+private TextView skipBtn;
 
     @Override
-    protected void onCreate(Bundle savedInstanceState) {
+    protected void onCreate(@Nullable Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_main);
+        initViews();
+        setListeners();
+    }
 
-        if (checkPlayServices()) {
-            setContentView(R.layout.activity_map);
+    private void initViews(){
+        skipBtn = (Button) findViewById(R.id.skipBtn);
+    }
 
-            SupportMapFragment mapFragment =
-                    (SupportMapFragment) getSupportFragmentManager().findFragmentById(R.id.map);
-            mapFragment.getMapAsync(this);
-
-            mLocationClient = new GoogleApiClient.Builder(this)
-                    .addApi(LocationServices.API)
-                    .addConnectionCallbacks(this)
-                    .addOnConnectionFailedListener(this)
-                    .build();
-
-            mLocationClient.connect();
-
+    private void setListeners(){
+        skipBtn.setOnClickListener(this);
+    }
+    @Override
+    public void onClick(View view) {
+        switch (view.getId()){
+            case R.id.skipBtn:
+                Intent skipIntent = new Intent(MainActivity.this, PointOfInterestActivity.class);
+                startActivity(skipIntent);
         }
     }
-
-
-    private boolean checkPlayServices() {
-        GoogleApiAvailability googleAPI = GoogleApiAvailability.getInstance();
-        int result = googleAPI.isGooglePlayServicesAvailable(this);
-        if (result != ConnectionResult.SUCCESS) {
-            if (googleAPI.isUserResolvableError(result)) {
-                googleAPI.getErrorDialog(this, result,
-                        PLAY_SERVICES_RESOLUTION_REQUEST).show();
-            }
-            return false;
-        }
-        return true;
-    }
-
-
-    @Override
-    public void onMapReady(GoogleMap googleMap) {
-        mMap = googleMap;
-        LatLng googleHQ = new LatLng(40.741815, -74.004230);
-        mMap.addMarker(new MarkerOptions().position(googleHQ).title("Marker at Google HQ"));
-        mMap.moveCamera(CameraUpdateFactory.newLatLng(googleHQ));
-
-        mMap.getUiSettings().setZoomControlsEnabled(true);
-        mMap.animateCamera(CameraUpdateFactory.zoomTo(15), 5000, null);
-    }
-
-
-    @Override
-    public void onConnected(@Nullable Bundle bundle) {
-        Toast.makeText(this, "Map working!!", Toast.LENGTH_SHORT).show();
-    }
-
-    @Override
-    public void onConnectionSuspended(int i) {
-        Toast.makeText(this, "Map creation interrupted", Toast.LENGTH_SHORT).show();
-    }
-
-    @Override
-    public void onConnectionFailed(@NonNull ConnectionResult connectionResult) {
-        Toast.makeText(this, "Map not connected!!", Toast.LENGTH_SHORT).show();
-    }
-
 }
