+++ app/src/main/java/nyc/c4q/helenchan/makinghistory/SplashActivity.java
@@ -1,0 +1,13 @@
+package nyc.c4q.helenchan.makinghistory;
+
+import android.support.v7.app.AppCompatActivity;
+import android.os.Bundle;
+
+public class SplashActivity extends AppCompatActivity {
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_splash);
+    }
+}
