+++ app/src/main/res/layout/activity_main.xml
@@ -1,17 +1,42 @@
 <?xml version="1.0" encoding="utf-8"?>
-<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
+<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools"
     android:id="@+id/activity_main"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    android:paddingBottom="@dimen/activity_vertical_margin"
-    android:paddingLeft="@dimen/activity_horizontal_margin"
-    android:paddingRight="@dimen/activity_horizontal_margin"
-    android:paddingTop="@dimen/activity_vertical_margin"
     tools:context="nyc.c4q.helenchan.makinghistory.MainActivity">
 
-    <TextView
+    <android.support.v7.widget.CardView
+        android:id="@+id/main_card"
+        android:layout_width="200dp"
+        android:layout_height="300dp"
+        android:layout_gravity="center_vertical"
+        android:layout_marginLeft="75dp">
+
+        <ImageView
+            android:id="@+id/main_pic_iv"
+            android:layout_width="@dimen/main_picture"
+            android:layout_height="@dimen/main_picture"
+            android:layout_gravity="center_horizontal"
+            android:background="@color/primary_light"
+            android:paddingTop="@dimen/main_pic_margin" />
+
+        <TextView
+            android:id="@+id/main_pic_tv"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:layout_gravity="bottom"
+            android:layout_margin="@dimen/main_pic_margin"
+            android:ellipsize="end"
+            android:maxLines="6"
+            android:text="@string/samuel_ipsum" />
+    </android.support.v7.widget.CardView>
+
+    <Button
+        android:id="@+id/skipBtn"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:text="Hello World!" />
-</RelativeLayout>
+        android:layout_gravity="bottom"
+        android:text="@string/skip" />
+
+</LinearLayout>
